ThanhLD: update stamina when in-app payment successfully

+++ Src/MangaApp/MangaApp/Common/StaminaConfig.h
@@ -11,7 +11,7 @@
 @interface StaminaConfig : NSObject
 
 @property (nonatomic, assign) NSInteger stamina;
-@property (nonatomic, assign) NSInteger maxStamina;
+@property (nonatomic, assign) float maxStamina;
 @property (nonatomic, strong) NSMutableArray *chapTrackList;
 @property (nonatomic, strong) NSString *chapName;
